pyproject.toml
[project]
name = "aidomaincontext"
version = "0.1.0"
description = "Enterprise RAG system for unified company knowledge search"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.6",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "openai>=1.60.0",
    "anthropic>=0.44.0",
    "unstructured[all-docs]>=0.16.0",
    "tiktoken>=0.8.0",
    "structlog>=24.4.0",
    "arq>=0.26.1",
    "apscheduler>=3.10.4",
    "sentence-transformers>=3.3.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
    "sse-starlette>=2.2.0",
    "google-genai>=1.67.0",
    "cryptography>=46.0.5",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aidomaincontext"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py313"
line-length = 100
